Split the substitution e2e tests out of proxy-tls.test.ts
proxy-tls.test.ts is described as covering the cert-trust, CONNECT and
streaming mechanics that plain-http tests can't reach, but substitution
behavior had grown to 8 of its 15 tests and roughly half its lines. Those
tests live there only because the proxy refuses to inject into a
cleartext connection, so anything that substitutes a secret needs the
MITM harness. That's a reason to share the harness, not to file the
tests under TLS.

Extract the harness (stub upstream + CONNECT tunnel + raw
send/read) into mitm-test-harness.ts, following the existing
env-graph test/helpers convention, and move the substitution tests to
proxy-substitution.test.ts. proxy-tls.test.ts keeps the transport and
TLS invariants plus the two policy tests. No test bodies changed; the
DNS-poison case now mints its wrong-host leaf from harness.upstreamCa().

The harness is not a build entry, so it stays out of dist.
